style: form construction
