+++ styles/default.css
@@ -30,7 +30,6 @@ code {
     padding: 10px 15px;
     background-color: #6d4aff;
     color: white;
-    margin: 5px;
     border-radius: 4px;
     cursor: pointer;
     text-decoration: none;
@@ -43,7 +42,6 @@ code {
     padding: 10px 15px;
     background-color: #4f31ce;
     color: white;
-    margin: 5px;
     border-radius: 4px;
     cursor: pointer;
     text-decoration: none;
@@ -56,7 +54,6 @@ code {
     padding: 10px 15px;
     background-color: #4f31ce;
     color: white;
-    margin: 5px;
     border-radius: 4px;
     cursor: pointer;
     text-decoration: none;
